Clear buffer after multicastAlarm_ric received
Fix for networks that mix normal alarms with multicastAlarms (delimiter needs to be deactivated).
If the delimiter is deactivated, RICs in the buffer will be deleted only if they are older than multicastAlarm_ignore_time. This could cause a mixture of RICs of the current and the privious multicastAlarm. This problem can be fixed, by clearing the buffer after the multicastAlarm sequence has finished.